% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosMXOutput.R
\name{CosMXOutput}
\alias{CosMXOutput}
\title{CosMX output from mouse brain}
\usage{
CosMXOutput(dataset = datasets, file_path = ".", force = FALSE, verbose = TRUE)
}
\arguments{
\item{dataset}{Can only be "subset1".}

\item{file_path}{Path to save downloaded files for the \code{*Output}
functions which don't return an SFE object.}

\item{force}{Logical, whether to force redownload if the files are already
present. Defaults to \code{FALSE}.}

\item{verbose}{Whether to display progress of download.}
}
\value{
Path to the tarball containing the output directory.
}
\description{
This is a small subset of the CosMX output from the mouse quarter brain
downloaded from the
\href{https://nanostring.com/resources/coronal-hippocampus-and-cortex-basic-data-files/}{Nanostring
website}. When downloaded, the files are in the output format of CosMX, not
an SFE object. The purpose of this dataset is to demonstrate and test
\code{readCosMX} in SFE.
}
