% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat.R
\name{SeuratTestData}
\alias{SeuratTestData}
\title{Seurat objects to test coercion function}
\usage{
SeuratTestData(
  dataset = datasets,
  file_path = ".",
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dataset}{Which dataset to use. Description of the datasets:
\describe{
\item{Visium}{From \code{SeuratData} (ie \code{stxBrain.SeuratData}), subsetted to keep first 50 genes}
\item{VisiumHD}{Visium HD mouse brain data from the \href{https://www.10xgenomics.com/datasets/visium-hd-cytassist-gene-expression-libraries-of-mouse-brain-he}{10X website},
with the first 50 genes in 8 um and 16 um bins}
\item{Vizgen}{Same dataset in \code{VizgenOutput} with \code{dataset = "hdf5"}}
\item{VizgenMulti}{Same as in \code{Vizgen} but with a subset of the data used
as if it's another sample to test the coercion function for multiple samples}
\item{Xenium}{Same as in \code{XeniumOutput}}
\item{XeniumMulti}{Same as in \code{Xenium} but with a subset of the data used
as if it's another sample to test the coercion function for multiple samples}
}}

\item{file_path}{Path to save downloaded files for the \code{*Output}
functions which don't return an SFE object.}

\item{force}{Logical, whether to force redownload if the files are already
present. Defaults to \code{FALSE}.}

\item{verbose}{Whether to display progress of download.}
}
\value{
A Seurat object
}
\description{
Some of the other test datasets in this package are made as Seurat object to
unit test function to convert Seurat objects to SFE.
}
