% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_info.R
\docType{package}
\name{SFEData}
\alias{SFEData}
\title{Example SpatialFeatureExperiment datasets}
\description{
Example spatial transcriptomics datasets with
\href{https://r-spatial.github.io/sf/}{Simple Features} annotations as
\code{SpatialFeatureExperiment} objects. In the first version, the only full
dataset is a mouse skeletal muscle Visium dataset 2 days after notexin injury
from 'Large-scale integration of single-cell transcriptomic data captures
transitional progenitor states in mouse skeletal muscle regeneration'. More
datasets will be added in later versions.
}
\section{Datasets}{


Full Visium dataset of the first time point, including spots outside
tissue.

Small subset of the full dataset for function examples.

A second small subset of the full dataset with a different \code{sample_id}
used for function examples involving multiple samples.
}

